refactor(site): deduplicate spend components
