+++ site/src/pages/AISettingsPage/SpendPage/components/LimitsTab/UserOverridesSection.tsx
@@ -12,6 +12,7 @@ import {
 	DialogHeader,
 	DialogTitle,
 } from "#/components/Dialog/Dialog";
+import { ConfirmDeleteDialog } from "#/components/Dialogs/ConfirmDeleteDialog/ConfirmDeleteDialog";
 import {
 	DropdownMenu,
 	DropdownMenuContent,
@@ -38,7 +39,6 @@ import {
 	isPositiveFiniteDollarAmount,
 } from "#/utils/currency";
 import { paginateItems } from "#/utils/paginateItems";
-import { ConfirmDeleteDialog } from "../ConfirmDeleteDialog";
 import { SectionHeader } from "../SectionHeader";
 
 const USER_OVERRIDES_PAGE_SIZE = 10;
